requirements.txt
h5py==2.7.1
imageio==2.6.1
imgaug==0.2.7
imutils==0.5.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.0.0
sklearn==0.0
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
matplotlib
